+++ packages/vee-validate/src/types/forms.ts
@@ -89,6 +89,9 @@ export interface PathState<TValue = unknown> {
   type: InputType;
   multiple: boolean;
   fieldsCount: number;
+  __flags: {
+    pendingUnmount: Record<string, boolean>;
+  };
   validate?: FieldValidator;
 }
 
@@ -228,6 +231,7 @@ export interface PrivateFormContext<TValues extends GenericObject = GenericObjec
   errors: ComputedRef<FormErrors<TValues>>;
   meta: ComputedRef<FormMeta<TValues>>;
   isSubmitting: Ref<boolean>;
+  isValidating: Ref<boolean>;
   keepValuesOnUnmount: MaybeRef<boolean>;
   validateSchema?: (mode: SchemaValidationMode) => Promise<FormValidationResult<TValues, TOutput>>;
   validate(opts?: Partial<ValidationOptions>): Promise<FormValidationResult<TValues, TOutput>>;
@@ -246,8 +250,9 @@ export interface PrivateFormContext<TValues extends GenericObject = GenericObjec
   ): PathState<PathValue<TValues, TPath>>;
   getPathState<TPath extends Path<TValues>>(path: TPath): PathState<PathValue<TValues, TPath>> | undefined;
   getAllPathStates(): PathState[];
-  removePathState<TPath extends Path<TValues>>(path: TPath): void;
+  removePathState<TPath extends Path<TValues>>(path: TPath, id: number): void;
   unsetPathValue<TPath extends Path<TValues>>(path: TPath): void;
+  markForUnmount(path: string): void;
 }
 
 export interface BaseComponentBinds<TValue = unknown> {
@@ -258,7 +263,7 @@ export interface BaseComponentBinds<TValue = unknown> {
 
 export type PublicPathState<TValue = unknown> = Omit<
   PathState<TValue>,
-  'bails' | 'label' | 'multiple' | 'fieldsCount' | 'validate' | 'id' | 'type'
+  'bails' | 'label' | 'multiple' | 'fieldsCount' | 'validate' | 'id' | 'type' | '__flags'
 >;
 
 export interface ComponentBindsConfig<TValue = unknown, TExtraProps extends GenericObject = GenericObject> {
@@ -313,6 +318,7 @@ export interface FormContext<TValues extends GenericObject = GenericObject, TOut
     | 'setFieldInitialValue'
     | 'unsetInitialValue'
     | 'fieldArrays'
+    | 'markForUnmount'
     | 'keepValuesOnUnmount'
   > {
   handleReset: () => void;