+++ packages/vee-validate/src/useFieldArray.ts
@@ -120,6 +120,7 @@ export function useFieldArray<TValue = unknown>(arrayPath: MaybeRef<string>): Fi
     const newValue = [...pathValue];
     newValue.splice(idx, 1);
     const fieldPath = pathName + `[${idx}]`;
+    form.markForUnmount(fieldPath);
     form.unsetInitialValue(fieldPath);
     setInPath(form.values, pathName, newValue);
     fields.value.splice(idx, 1);