+++ packages/nuxt/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @vee-validate/nuxt
 
+## 4.9.3
+
+### Patch Changes
+
+- Updated dependencies [09d5596b]
+- Updated dependencies [9bfbfaaf]
+- Updated dependencies [48b45d91]
+  - vee-validate@4.9.3
+
+## 4.9.2
+
+### Patch Changes
+
+- Updated dependencies [31090e0d]
+- Updated dependencies [9046308b]
+- Updated dependencies [fe322a07]
+  - vee-validate@4.9.2
+
 ## 4.9.1
 
 ### Patch Changes